+++ tests/test_multithread.py
@@ -467,24 +467,33 @@ def test_modify_acl_entries_recursive(azure):
     with setup_tree(azure):
         acluser = AZURE_ACL_TEST_APPID
 
-        def check_acl_perms(path, permission):
+        def check_acl_perms(path, permission, isdir=False):
             current_acl = azure.get_acl_status(path)
             acl_user_entry = [s for s in current_acl['entries'] if acluser in s]
-            assert len(acl_user_entry) == 1
+            if isdir:
+                assert len(acl_user_entry) == 2
+            else:
+                assert len(acl_user_entry) == 1
             assert acl_user_entry[0].split(':')[-1] == permission
 
         files = list(azure.walk(test_dir))
         directories = list(set([x[0] for x in map(os.path.split, files)]))
 
         permission = "---"
-        azure.modify_acl_entries(test_dir, acl_spec="user:"+acluser+":"+permission, recursive=True)
-        for path in files+directories:
-            check_acl_perms(path, permission)
+        azure.modify_acl_entries(test_dir, acl_spec="default:user:"+acluser+":"+permission+",user:"+acluser+":"+permission, recursive=True)
+        for path in files:
+            check_acl_perms(path, permission, False)
+
+        for path in directories:
+            check_acl_perms(path, permission, True)
 
         permission = "rwx"
-        azure.modify_acl_entries(test_dir, acl_spec="user:"+acluser+":"+permission, recursive=True)
-        for path in files+directories:
-            check_acl_perms(path, permission)
+        azure.modify_acl_entries(test_dir, acl_spec="default:user:"+acluser+":"+permission+",user:"+acluser+":"+permission, recursive=True)
+        for path in files:
+            check_acl_perms(path, permission, False)
+
+        for path in directories:
+            check_acl_perms(path, permission, True)
 
 
 @my_vcr.use_cassette